refactor(frontend): extract useClampOverflow into its own hook file
